dataflow-expression: rename post_lookup.rs
PostLookup has been extended into a more generic ReaderProcessing struct
that also includes pre-insertion and will soon be further extended, so
rename the file to match the main struct name.
